GlobalThreadCount uint64_t type

+++ ocl.h
@@ -14,7 +14,7 @@ typedef struct __clState {
   cl_mem CLbuffer0;
   cl_mem MidstateBuf;
   cl_mem BranchBuffer[6];
-  size_t GlobalThreadCount;
+  uint64_t GlobalThreadCount;
   cl_mem padbuffer8;
   cl_mem buffer1;
   cl_mem buffer2;
